P1.3c: Split user_group handler into per-resource files (#37)
Issue #6. internal/handlers/user_group.go was 760 LOC of uniform thin-handler
boilerplate (decode -> CanManage* check -> service call -> encode) mixing two
REST resources. The UserGroupService already owns all logic, so this split is
pure file-size partitioning within package handlers.

Method receivers stay on *UserGroupHandler, so router and bootstrap wiring is
unchanged.

- user_group.go (760 -> 43 LOC): struct, constructor, HandleUsers/HandleGroups dispatchers
- users.go (NEW, 295 LOC): user CRUD + DTOs
- groups.go (NEW, 295 LOC): group CRUD + DTOs
- group_membership.go (NEW, 142 LOC): AddUserToGroup, RemoveUserFromGroup + DTO
- route_assignment.go (MOD): picks up CreateRouteAssignmentRequest (misplaced
  in user_group.go, used only here)

Out of scope:
- Repeated "extract id from URL path" + permission-check boilerplate -> middleware
  candidates for Phase 2 controller-runtime migration
- pkg/services/services/user_group.go size unaudited

Closes #6
